+++ mkchromecast.py
@@ -5,7 +5,7 @@
 from mkchromecast.audiodevices import *
 from mkchromecast.cast import *
 from mkchromecast.terminate import *
-import os.path
+import os.path, time
 
 import atexit
 
@@ -44,8 +44,6 @@ if args.tray == False:
         cc.get_cc()
         cc.play_cast()
 
-    print('Ctrl-C to kill the application at any time')
-    print('')
 
     def terminateapp():
         cc.stop_cast()
@@ -55,9 +53,42 @@ if args.tray == False:
         return
 
     try:
-        input()
-    except KeyboardInterrupt:
-        atexit.register(terminateapp)
+        volumearg = mkchromecast.__init__.volumearg
+    except AttributeError:
+        volumearg = False
+
+    if volumearg == True:
+        from getch import getch, pause
+
+        print('Controls:')
+        print('========')
+        print('')
+        print('Volume up: u')
+        print('Volume up: d')
+        print('Quit the application: q')
+        print('')
+        try:
+            while(True):
+                key = getch()
+                if(key == 'u'):
+                    print('Increasing volume...')
+                    cc.volume_up()
+                elif(key == 'd'):
+                    print('Decreasing volume...')
+                    cc.volume_down()
+                elif(key == 'q'):
+                    print('Quitting application...')
+                    terminateapp()
+        except KeyboardInterrupt:
+            terminateapp()
+
+    else:
+        print('Ctrl-C to kill the application at any time')
+        print('')
+        try:
+            input()
+        except KeyboardInterrupt:
+            atexit.register(terminateapp)
 else:
     import mkchromecast.systray
     checkmktmp()